Genome Feature Type Definitions

A dataset containing different types of gene and genome features along with their Sequence Ontology (SO) identifiers and definitions.

Usage

feature_type_definitions

Format

A tibble with 71 rows and 3 variables:

feature_type

Character. The type of gene or genome feature.

so_id

Character. The Sequence Ontology identifier associated with the feature type.

definition

Character. The definition of the feature type.

Source

The table in https://www.informatics.jax.org/userhelp/GENE_feature_types_help.shtml and a few other terms found in MGI reports.

Examples

print(feature_type_definitions, n = Inf)
#> # A tibble: 71 × 3
#>    feature_type                          so_id      definition                  
#>    <chr>                                 <chr>      <chr>                       
#>  1 gene                                  SO:0000704 A region (or regions) that …
#>  2 protein coding gene                   SO:0001217 A gene that produces at lea…
#>  3 gene segment                          SO:3000000 A gene component region whi…
#>  4 unclassified gene                     NA         A region of the genome asso…
#>  5 heritable phenotypic marker           SO:0001500 A biological region charact…
#>  6 other feature types                   NA         MGI markers that are not cl…
#>  7 QTL                                   SO:0000771 A quantitative trait locus …
#>  8 transgene                             SO:0000902 A gene that has been transf…
#>  9 non-coding RNA gene                   SO:0001263 A gene that produces an RNA…
#> 10 lncRNA gene                           SO:0001877 A gene that encodes a non-c…
#> 11 antisense lncRNA gene                 SO:0002182 A gene that encodes a non-c…
#> 12 lincRNA gene                          SO:0001641 A gene that encodes large i…
#> 13 sense intronic lncRNA gene            SO:0002184 A gene that encodes a sense…
#> 14 sense overlapping lncRNA gene         SO:0002183 A gene that encodes a sense…
#> 15 bidirectional promoter lncRNA gene    SO:0002185 A non-coding locus that ori…
#> 16 rRNA gene                             SO:0001637 A gene that encodes ribosom…
#> 17 tRNA gene                             SO:0001272 A gene that encodes Transfe…
#> 18 snRNA gene                            SO:0001268 A gene that encodes a Small…
#> 19 snoRNA gene                           SO:0001267 A gene that encodes for Sma…
#> 20 miRNA gene                            SO:0001265 A gene that encodes for mic…
#> 21 scRNA gene                            SO:0001266 A gene that encodes for Sma…
#> 22 SRP RNA gene                          SO:0000590 A gene that encodes the sig…
#> 23 RNase P RNA gene                      SO:0001639 A gene that encodes RNase P…
#> 24 RNase MRP RNA gene                    SO:0001640 A gene that encodes RNase M…
#> 25 telomerase RNA gene                   SO:0001643 A non-coding RNA gene, the …
#> 26 unclassified non-coding RNA gene      NA         A non-coding RNA gene not c…
#> 27 ribozyme gene                         SO:0002181 A gene that encodes an RNA …
#> 28 complex/cluster/region                NA         A group of linked markers c…
#> 29 cytogenetic marker                    NA         A structure within a chromo…
#> 30 chromosomal deletion                  SO:1000029 An incomplete chromosome.   
#> 31 insertion                             SO:0000667 The sequence of one or more…
#> 32 chromosomal inversion                 SO:1000030 An interchromosomal mutatio…
#> 33 Robertsonian fusion                   SO:1000043 A non reciprocal translocat…
#> 34 reciprocal chromosomal translocation  SO:1000048 A chromosomal translocation…
#> 35 chromosomal translocation             SO:1000044 An interchromosomal mutatio…
#> 36 chromosomal duplication               SO:1000037 An extra chromosome.        
#> 37 chromosomal transposition             SO:0000453 A chromosome structure vari…
#> 38 unclassified cytogenetic marker       NA         A cytogenetic marker not cl…
#> 39 BAC/YAC end                           NA         A region of sequence from t…
#> 40 BAC end                               SO:0000999 A region of sequence from t…
#> 41 YAC end                               SO:0001498 A region of sequence from t…
#> 42 PAC end                               SO:0001480 A region of sequence from t…
#> 43 other genome feature                  NA         A region of the genome asso…
#> 44 retrotransposon                       SO:0000180 A transposable element that…
#> 45 telomere                              SO:0000624 A specific structure at the…
#> 46 minisatellite                         SO:0000643 A repeat region containing …
#> 47 unclassified other genome feature     NA         A genome feature that canno…
#> 48 endogenous retroviral region          SO:0000903 A region derived from viral…
#> 49 mutation defined region               NA         A genomic region, containin…
#> 50 CpG island                            SO:0000307 Regions of a few hundred to…
#> 51 promoter                              SO:0000167 A regulatory region compose…
#> 52 promoter flanking region              SO:0001952 Transcription factor bindin…
#> 53 TSS region                            SO:0001240 The region of a gene from t…
#> 54 enhancer                              SO:0000165 A cis-acting sequence that …
#> 55 CTCF binding site                     SO:0001974 A transcription factor bind…
#> 56 transcription factor binding site     SO:0000235 A DNA site where a transcri…
#> 57 open chromatin region                 SO:0001747 Regions of spaced out histo…
#> 58 DNA segment                           NA         A region of the genome asso…
#> 59 pseudogenic region                    SO:0000462 A non-functional descendant…
#> 60 pseudogene                            SO:0000336 A sequence that closely res…
#> 61 pseudogenic gene segment              SO:0001741 A recombinational unit of a…
#> 62 polymorphic pseudogene                SO:0001841 Pseudogene owing to a SNP/D…
#> 63 imprinting control region             SO:0002191 A regulatory region that co…
#> 64 intronic regulatory region            SO:0001492 A regulatory region that is…
#> 65 silencer                              SO:0000625 A regulatory region which u…
#> 66 locus control region                  SO:0000037 A DNA region that includes …
#> 67 insulator                             SO:0000627 A regulatory region that 1)…
#> 68 response element                      SO:0002205 A regulatory element that a…
#> 69 origin of replication                 SO:0000296 A region of nucleic acid fr…
#> 70 transcriptional cis regulatory region SO:0001055 A regulatory_region that mo…
#> 71 TSS cluster                           SO:0001915 A region defined by a clust…