Read an MGI report

read_report() imports data from an MGI report into R as a tidy data set.

You may call this function in two alternative ways:

Using report_key: this is the easiest approach. A report key maps to a report currently hosted at MGI, e.g. read_report("marker_list2") reads MRK_List2.rpt directly from MGI server into R. See Supported Reports below for options.

Using report_file and report_type: this approach is more flexible as you can read directly from a file or URL.

Supported Reports

The set of currently supported reports:

reports
#> # A tibble: 13 x 4
#>    report_key                 report_file              report_type   report_name
#>    <chr>                      <chr>                    <chr>         <chr>      
#>  1 marker_list1               MRK_List1.rpt            MRK_List1     Mouse Gene~
#>  2 marker_list2               MRK_List2.rpt            MRK_List2     Mouse Gene~
#>  3 marker_coordinates         MGI_MRK_Coord.rpt        MGI_MRK_Coord MGI Marker~
#>  4 gene_model_coordinates     MGI_Gene_Model_Coord.rpt MGI_Gene_Mod~ MGI Gene M~
#>  5 sequence_coordinates       MGI_GTGUP.gff            MGI_GTGUP     MGI Sequen~
#>  6 genbank_refseq_ensembl_ids MRK_Sequence.rpt         MRK_Sequence  MGI Marker~
#>  7 swiss_trembl_ids           MRK_SwissProt_TrEMBL.rpt MRK_SwissPro~ MGI Marker~
#>  8 swiss_prot_ids             MRK_SwissProt.rpt        MRK_SwissProt MGI Marker~
#>  9 gene_trap_ids              MRK_GeneTrap.rpt         MRK_GeneTrap  MGI Marker~
#> 10 ensembl_ids                MRK_ENSEMBL.rpt          MRK_ENSEMBL   MGI Marker~
#> 11 biotype_conflicts          MGI_BioTypeConflict.rpt  MGI_BioTypeC~ MGI Marker~
#> 12 primers                    PRB_PrimerSeq.rpt        PRB_PrimerSeq MGI Marker~
#> 13 interpro_domains           MGI_InterProDomains.rpt  MGI_InterPro~ InterPro d~

Usage

read_report(
  report_key = NULL,
  report_file = NULL,
  report_type = NULL,
  n_max = Inf
)

Arguments

report_key

A character vector. A key used to uniquely refer to an MGI report.

report_file

A character vector. The file path or URL to an MGI report file.

report_type

A character vector. The type of an MGI report.

n_max

Maximum number of lines to read.

Value

A tibble with report data in tidy format. The set of variables is dependent on the specific report requested:

For "marker_list1", see vignette("marker_list1").

For "marker_list2", see vignette("marker_list2").

For "marker_coordinates", see vignette("marker_coordinates").

For "gene_model_coordinates", see vignette("gene_model_coordinates").

For "sequence_coordinates", see vignette("sequence_coordinates").

For "genbank_refseq_ensembl_ids", see vignette("genbank_refseq_ensembl_ids").

For "swiss_trembl_ids", see vignette("swiss_trembl_ids").

For "swiss_prot_ids", see vignette("swiss_prot_ids").

For "gene_trap_ids", see vignette("gene_trap_ids").

For "ensembl_ids", see vignette("ensembl_ids").

For "biotype_conflicts", see vignette("biotype_conflicts").

For "primers", see vignette("primers").

For "interpro_domains", see vignette("interpro_domains").