Supported MGI reports

reports is a data set of supported MGI reports, meaning reports that {mgi.report.reader} can currently read into R. To browse all reports made available by MGI visit https://www.informatics.jax.org/downloads/reports.

Usage

reports

Format

A tibble of 4 variables:

report_key

A string key used to uniquely refer to an MGI report, which is only meaningful within the context of the {mgi.report.reader}.

report_file

MGI report file name as hosted at https://www.informatics.jax.org/downloads/reports .

report_type

MGI report type. The type is used internally to find the appropriate reader for parsing, and is only meaningful within the context of {mgi.report.reader}.

report_name

MGI report name. Report names are taken from https://www.informatics.jax.org/downloads/reports/index.html .

Examples

reports
#> # A tibble: 13 × 4
#>    report_key                 report_file              report_type   report_name
#>    <chr>                      <chr>                    <chr>         <chr>      
#>  1 marker_list1               MRK_List1.rpt            MRK_List1     Mouse Gene…
#>  2 marker_list2               MRK_List2.rpt            MRK_List2     Mouse Gene…
#>  3 marker_coordinates         MGI_MRK_Coord.rpt        MGI_MRK_Coord MGI Marker…
#>  4 gene_model_coordinates     MGI_Gene_Model_Coord.rpt MGI_Gene_Mod… MGI Gene M…
#>  5 sequence_coordinates       MGI_GTGUP.gff            MGI_GTGUP     MGI Sequen…
#>  6 genbank_refseq_ensembl_ids MRK_Sequence.rpt         MRK_Sequence  MGI Marker…
#>  7 swiss_trembl_ids           MRK_SwissProt_TrEMBL.rpt MRK_SwissPro… MGI Marker…
#>  8 swiss_prot_ids             MRK_SwissProt.rpt        MRK_SwissProt MGI Marker…
#>  9 gene_trap_ids              MRK_GeneTrap.rpt         MRK_GeneTrap  MGI Marker…
#> 10 ensembl_ids                MRK_ENSEMBL.rpt          MRK_ENSEMBL   MGI Marker…
#> 11 biotype_conflicts          MGI_BioTypeConflict.rpt  MGI_BioTypeC… MGI Marker…
#> 12 primers                    PRB_PrimerSeq.rpt        PRB_PrimerSeq MGI Marker…
#> 13 interpro_domains           MGI_InterProDomains.rpt  MGI_InterPro… InterPro d…